# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IObenkyo api",
        "version": "1.0"
    },
    "paths": {
...
        "/api/habits/follow-up/{habitId}": {
            "get": {
                "tags": [
                    "Habits follow up"
                ],
                "summary": "Lists all follow ups on a habit",
                "operationId": "App\\Http\\Controllers\\HabitFollowUpController::followUpList",
                "responses": {
                    "200": {
                        "description": "List of follow ups",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/HabitFollowUp"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Habits follow up"
                ],
                "summary": "Allows to mark a follow up for the given day",
                "operationId": "App\\Http\\Controllers\\HabitFollowUpController::followUpMark",
                "parameters": [
                    {
                        "name": "apply_date",
                        "in": "query",
                        "description": "Date to mark the habit as followed (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "story",
                        "in": "query",
                        "description": "A short description about the habit",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "counter",
                        "in": "query",
                        "description": "A value to mark the times completed an habit",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "remove",
                        "in": "query",
                        "description": "If the follow up should be removed",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "update",
                        "in": "query",
                        "description": "If the follow up should be updated",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "accomplished",
                        "in": "query",
                        "description": "If the follow up should be mark as completed",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The updated or saved follow up",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HabitFollowUp"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/habits/daily-follow-up/{dateStr}": {
            "get": {
                "tags": [
                    "Habits follow up"
                ],
                "summary": "Lists the habits for a given day",
                "operationId": "App\\Http\\Controllers\\HabitFollowUpController::dailyFollowUp",
                "parameters": [
                    {
                        "name": "dateStr",
                        "in": "path",
                        "description": "The day to search for follow ups",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Follow ups for given name",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/HabitFollowUp"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Habit": {
                "properties": {
                    "id": {
                        "description": "",
                        "type": "integer"
                    },
                    "title": {
                        "description": "",
                        "type": "string"
                    },
                    "description": {
                        "description": "",
                        "type": "string"
                    },
                    "start": {
                        "description": "Date where the habit should start",
                        "type": "string",
                        "default": "YYYY-MM-DD HH:mm:ss"
                    },
                    "goal_date": {
                        "description": "",
                        "type": "string",
                        "default": "YYYY-MM-DD HH:mm:ss"
                    },
                    "streak_count": {
                        "description": "Current streak count",
                        "type": "integer"
                    },
                    "streak_goal": {
                        "description": "Amount to be used as goal",
                        "type": "integer"
                    },
                    "category": {
                        "$ref": "#/components/schemas/HabitCategory"
                    },
                    "is_counter": {
                        "description": "If the habit should be accomplished by completing an amount of times",
                        "type": "boolean"
                    },
                    "counter_goal": {
                        "description": "Amount of times to complete the goal a single day",
                        "type": "integer"
                    },
                    "max_streak": {
                        "description": "The user max streak in a row",
                        "type": "integer"
                    },
                    "should_keep": {
                        "description": "If the habit its something that benefits the user",
                        "type": "boolean"
                    },
                    "should_avoid": {
                        "description": "If the habit its something that demages the user",
                        "type": "boolean"
                    },
                    "followUps": {
                        "description": "",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/HabitFollowUp"
                        }
                    },
                    "created_at": {
                        "description": "Created date",
                        "type": "string",
                        "default": "YYYY-MM-DD HH:mm:ss"
                    },
                    "updated_at": {
                        "description": "Updated date",
                        "type": "string",
                        "default": "YYYY-MM-DD HH:mm:ss"
                    }
                },
                "type": "object"
            },
            "HabitCategory": {
                "properties": {
                    "id": {
                        "description": "",
                        "type": "integer"
                    },
                    "name": {
                        "description": "",
                        "type": "string"
                    },
                    "icon": {
                        "description": "",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Created date",
                        "type": "string",
                        "default": "YYYY-MM-DD HH:mm:ss"
                    },
                    "updated_at": {
                        "description": "Updated date",
                        "type": "string",
                        "default": "YYYY-MM-DD HH:mm:ss"
                    }
                },
                "type": "object"
            },
            "HabitFollowUp": {
                "properties": {
                    "id": {
                        "description": "",
                        "type": "integer"
                    },
                    "apply_date": {
                        "description": "",
                        "type": "string",
                        "default": "YYYY-MM-DD HH:mm:ss"
                    },
                    "story": {
                        "description": "A description of what happened while accomplishing this habit",
                        "type": "string"
                    },
                    "accomplished": {
                        "description": "",
                        "type": "boolean"
                    },
                    "counter": {
                        "description": "Times completed by the user",
                        "type": "integer"
                    },
                    "counter_goal": {
                        "description": "Times to mark the goal as accomplished",
                        "type": "integer"
                    },
                    "is_counter": {
                        "description": "If the follow up is completed by counting",
                        "type": "boolean"
                    },
                    "habit": {
                        "$ref": "#/components/schemas/Habit"
                    },
                    "created_at": {
                        "description": "Created date",
                        "type": "string",
                        "default": "YYYY-MM-DD HH:mm:ss"
                    },
                    "updated_at": {
                        "description": "Updated date",
                        "type": "string",
                        "default": "YYYY-MM-DD HH:mm:ss"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "security": [
        []
    ]
}